+++ publish/CHANGELOG.md
@@ -7,10 +7,9 @@
   - Fixes custom dungeons not working.
   - Fixes the default door spawn change being 1 instead of 0 (too many doors appeared for some dungeons).
   - Fixes the field `paint` in the `expand_biomes.yaml` file not working when set to `0,0,0,0`.
+  - Fixes the health data not being ignored for locations with random damage.
   - Removes creator information automatically from blueprints and data values (so that things don't appear player built).
   - Improves the performance of the field `paint` in the `expand_biomes.yaml` file.
-  - TODO: Check that no build works.
-  - TODO: Check that random health works for blueprints (ignore health data if given).
   
 - v1.40
   - Adds blueprint dungeons rooms.